More float/double parsing fix in Postgres.
Since float/double value can be NaN, converters can't assume that N stands for NULL.
Check additional character for correct value.

While this is not really safest code - it would be better to read value in whole and use that for decision,
currently because of performance it's assumed that values returned from Postgres are correct.