Event queueing.
Allow fire and forget on events. Such events will be persisted with best effort (out of current transaction).

This can be used for logging events even in case of rollback or for fast event submission.

Allow null as stream result in Http server. Interpret this as 0 length without content type.