IPrincipal passed around. VS 2013 sln
IPrincipal is also registered into the container, but since it's not really a service but rather a security context, pass it around.

Added Revenj solution for VS 2013 (with reference to new SignalR)

TODO: currently object factory is passed to processing engine instead of service locator. They differ by service locator trying to resolve stuff not registered into the container.