[Feature] Visibility management improvements
A common use case in forms is the need to control the visibility of sections or controls via some predicate. So far, it was available only on the group level, and only with inline functions.

Expose a registry for utilities for visibility that can be used straight from DSL after registration as `this.context.Visibility.yourFunction`.

Furthermore, expose a `visible` property (either static or dynamic) on the fields/controls themselves. Manage its clearing of values when it's in a possibly invisible context, so that we do not have dead values lying around.

Tweak the license name to match the standard