[Feature] Standardise i18n
Make the I18n feature work on all titles and labels in the application, not just report titles and grid column titles.

Standardise it in a way that it will just look up anything that starts with the prefix `i18n:`, followed by some unique identifier. The handling of the identifier is managed by the application that implements/leverages revenj.

Update the documentation to reflect the change