Add example app + fixes
Create a small example application. This application sets up a basic
CRUD based on DSL-generated code and the newest version of Revenj to
showcase the basic development flow and the necessary boilerplate when
developing such an application, as well as a brief explanation in its
README file.

Fix several issues while at it:

* Fix an issue where a required field of an optional structure would still be required - it should inherit optionality
* Change the default Fields we need to register (special components for Numbers, LongText instead of Textarea, as generated)
* Fix typings for formField helper
* Fix component extensions - they should not require re-registration of default fields, just extra ones